pyproject.toml
[project]
name = "cloud-ai-centric-course"
version = "0.1.0"
description = "Curso: RAG, LangChain, LangGraph y GCP para construir agentes de IA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.13",
    "langchain-core>=1.2.12",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.4.1",
    "langgraph>=1.0.9",
    "langsmith>=0.7.6",
    "openai>=2.23.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "pydantic>=2.10.6",
    "tiktoken>=0.12.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_ai"]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.5",
    "ruff>=0.15.2",
]
